Convert GossipListenAddr to use env_config_socketaddr macro
Also consolidate with ListenCtlAddr under components/common/src/types/mod.rs
