+++ components/hab/src/cli.rs
@@ -5,7 +5,6 @@ use clap::{App,
 use habitat_common::{cli::{BINLINK_DIR_ENVVAR,
                            DEFAULT_BINLINK_DIR,
                            GOSSIP_DEFAULT_ADDR,
-                           GOSSIP_LISTEN_ADDRESS_ENVVAR,
                            LISTEN_CTL_DEFAULT_ADDR_STRING,
                            LISTEN_HTTP_ADDRESS_ENVVAR,
                            LISTEN_HTTP_DEFAULT_ADDR,
@@ -16,6 +15,7 @@ use habitat_common::{cli::{BINLINK_DIR_ENVVAR,
                            SHUTDOWN_TIMEOUT_DEFAULT},
                      types::{AutomateAuthToken,
                              EventStreamMetadata,
+                             GossipListenAddr,
                              ListenCtlAddr},
                      FeatureFlag};
 use habitat_core::{crypto::{keys::PairType,
