+++ components/hab/src/main.rs
@@ -47,8 +47,8 @@ use habitat_core::{crypto::{init,
                             keys::PairType,
                             BoxKeyPair,
                             SigKeyPair},
-                   env as henv,
-                   env::Config as EnvConfig,
+                   env::{self as henv,
+                         Config as _},
                    fs::{cache_analytics_path,
                         cache_artifact_path,
                         launcher_root_path},
